Finalización Servicios Inventarios

+++ bambu/src/main/java/com/panda/bambu/service/inventory/ArticleInventoryService.java
@@ -1,30 +1,44 @@
 package com.panda.bambu.service.inventory;
 
 import java.util.List;
+import java.util.TimerTask;
 
 import com.panda.bambu.model.article.Article;
 import com.panda.bambu.model.inventory.ArticleInventory;
 import com.panda.bambu.model.inventory.ArticleInventoryRepository;
+import com.panda.bambu.model.inventory.Balance;
+import com.panda.bambu.model.inventory.Entry;
+import com.panda.bambu.model.inventory.Inventory;
+import com.panda.bambu.model.inventory.Output;
 import com.panda.bambu.service.article.ArticleService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ArticleInventoryService {
+public class ArticleInventoryService extends TimerTask{
 
     @Autowired
     private ArticleInventoryRepository articleIRepository;
 
     @Autowired
-	ArticleService articleService;
+    ArticleService articleService;
+    
+    @Autowired
+    InventoryService inventoryService;
+    
+    @Autowired
+    EntryService entryService;
+    
+    @Autowired
+	OutputService outputService;
 
     public ArticleInventory findById(Long id){
         return articleIRepository.findById(id).get();
     }
         
-    public ArticleInventory findByCode(Article article){
-         return articleIRepository.findByCode(article.getCode());
+    public ArticleInventory findByArticle(Article article){
+         return articleIRepository.findByArticle(article);
     }
     
     public List<ArticleInventory> findAll(){
@@ -33,12 +47,12 @@ public class ArticleInventoryService {
 
     public Boolean create(Article article) {
        
-        if(articleIRepository.findByCode(article.getCode()) != null){
+        if(articleService.isArticleAlreadyPresent(article) == true){
             if(article !=null){
                 ArticleInventory newArticle = new ArticleInventory();
-                newArticle.setCode(article.getCode());
                 newArticle.setArticle(article);    
                 articleIRepository.save(newArticle);
+                newArticle.getInventories().add(inventoryService.create());
                 return true;            
             }
         }
@@ -49,56 +63,24 @@ public class ArticleInventoryService {
         
         Article article = articleService.findById(articleId);
 
-        if(articleIRepository.findByCode(article.getCode()) != null){
-            if(article !=null){
-                ArticleInventory newArticle = new ArticleInventory();
-                newArticle.setCode(article.getCode());
-                newArticle.setArticle(article);    
-                articleIRepository.save(newArticle);
-                return true;
-            }
-        }
-        return false;
-
-    }
-
-    public boolean create(String codeId){
-        
-        Article article = articleService.findByCode(codeId);
-        if(articleIRepository.findByCode(article.getCode()) != null){
-            if(article !=null){
-
-                ArticleInventory newArticle = new ArticleInventory();
-                newArticle.setCode(article.getCode());
-                newArticle.setArticle(article);    
-                articleIRepository.save(newArticle);            
-            }
+        if(article !=null){
+            ArticleInventory newArticle = new ArticleInventory();
+            newArticle.setArticle(article);    
+            articleIRepository.save(newArticle);
             return true;
         }
+        
         return false;
 
     }
     
-    public boolean delete(String code){
-          
-         ArticleInventory article = articleIRepository.findByCode(code);
-         if( article != null){
-             articleIRepository.delete(article);
-             return true;
-         } 
+    public boolean delete(ArticleInventory article){
 
-         return false;
-   }
-
-    public boolean delete(Article article){
-          
-         for (ArticleInventory element: articleIRepository.findAll()){
-              if(element.getArticle().equals(article)){
-                 articleIRepository.delete(element);
-                 return true;
-              }
-         }
-         return false;
+        if( article != null && articleIRepository.existsById(article.getId())){
+            articleIRepository.delete(article);
+            return true;
+        } 
+        return false;
     }
 
     public boolean delete(Long id){
@@ -111,7 +93,85 @@ public class ArticleInventoryService {
          return false;
     }
 
+    public boolean save(ArticleInventory article){
+
+        ArticleInventory articleF = articleIRepository.findByArticle(article.getArticle());
+        articleF = articleIRepository.findById(article.getId()).get();
+        if( articleF != null || articleIRepository.existsById(article.getId()) == true){
+            articleIRepository.save(article);
+            return true;
+        } 
+        return false;
+    }
+
     public void deleteAll(){
         articleIRepository.deleteAll();
     }
+
+    public boolean addEntry(ArticleInventory articleInventory, Entry entry){
+        
+          entry.setArticle(articleInventory.getArticle());
+          if(articleIRepository.existsById(articleInventory.getId()) == true && !entryService.create(entry)){
+             if(articleInventory.getInventories().isEmpty()){
+                articleInventory.getInventories().add(inventoryService.create());
+             }
+             entry.setArticle(articleInventory.getArticle());
+             entryService.create(entry);
+
+             articleInventory.getArticle().setUnitCost(inventoryService.addEntryInventory(articleInventory.getInventories().get(articleInventory.getInventories().size()-1), entry));
+             articleService.save(articleInventory.getArticle());
+             return true;
+          }
+
+          return false;
+    }
+
+    
+    public boolean addOuput(ArticleInventory articleInventory, Output output){
+        
+        output.setArticle(articleInventory.getArticle());
+        if(articleIRepository.existsById(articleInventory.getId()) == true && !outputService.create(output)){
+            if(articleInventory.getInventories().isEmpty()){
+                articleInventory.getInventories().add(inventoryService.create());
+                articleIRepository.save(articleInventory);
+             }
+             output.setArticle(articleInventory.getArticle());
+             outputService.create(output);   
+            
+             articleInventory.getArticle().setUnitCost(inventoryService.addOutputInventory(articleInventory.getInventories().get(articleInventory.getInventories().size()-1), output));
+             articleService.save(articleInventory.getArticle());
+             return true;
+        }
+
+        return false;
+    }
+
+      @Override
+      public void run() {
+            createInventory();
+      }
+
+      public void createInventory(){
+
+             for(ArticleInventory article: articleIRepository.findAll()){
+                 Inventory inventory = inventoryService.create();
+                 List<Inventory> inventories = article.getInventories();
+                 List<Balance> balances = inventories.get(inventories.size()-1).getBalances();
+                 Balance balance = balances.get(balances.size()-1);
+                 
+                 if(balance.getQuantity() == 0 && balance.getTotalCost() == 0){
+                    if(!inventoryService.addBalanceInventory(inventory,balance,article.getArticle())){
+                        article.getInventories().add(inventory);
+                        articleIRepository.save(article);
+                    }
+                  
+                 }
+                 else{
+                    article.getInventories().add(inventory);
+                    articleIRepository.save(article);
+                 }
+                 
+             }
+             
+      }
 }
